pyproject.toml
[project]
name = "python-mcp-server"
version = "1.0.0-beta"
description = "MCP server with Graphiti knowledge graph and RAG search for accurate AI responses"
requires-python = "==3.13.2"
dependencies = [
    "pydantic>=2.9.2",
    "mcp[cli]>=1.0.0",
    "graphiti-core>=0.3.0",
    "asyncpg>=0.29.0",
    "pgvector>=0.3.0",
    "pyyaml>=6.0.0",
    "starlette>=0.49.1",
    "openai>=1.54.0",
]

[project.scripts]
python-mcp-server = "python_mcp_server.__main__:main"

[tool.uv]
package = true
constraint-dependencies = [
    "cryptography>=46.0.7",   # CVE-2026-39892
    "pygments>=2.20.0",       # CVE-2026-4539
    "requests>=2.33.0",       # CVE-2026-25645
    "aiohttp>=3.13.4",        # 10 CVEs in 3.13.3
]

[dependency-groups]
dev = [
    "bandit>=1.8.0",
    "black[d]>=26.3.1",
    "mypy>=1.14.0",
    "poethepoet>=0.32.0",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.8.4",
    "pip-audit>=2.6.1",
    "pytest-cov>=6.0.0",
    "deptry>=0.23.0",
    "pip>=25.1.1",
]

[tool.poe.tasks]
depcheck = "deptry . --per-rule-ignores DEP004=pytest,DEP002=asyncpg|starlette"
format = "black ."
lint = "ruff check"
typecheck = "mypy  --install-types --non-interactive  --explicit-package-bases ."
audit-src = "bandit -r src/* -s B101,B608"
audit-packages = "pip-audit --ignore-vuln GHSA-4xh5-x5gv-qwph"
security = ["audit-src", "audit-packages"]
checks = ["depcheck", "format", "lint", "typecheck", "security"]
unit = "pytest -vv src/*"
integration = "pytest -vv tests/*"
cover = "pytest -vv --cov"
review = "claude 'run code-reviewer agent against diff for quality and elegance'"
commit = "bash -c 'uv run poe checks && uv run poe cover && uv run poe review && git add -A && git cz && git push'"

[tool.coverage.run]
omit = ["*/test_*.py", "tests/*"]
